# schemas/openapi-commercial-v1.in.yaml
openapi: "3.0.3"
info:
  title: AgentSkeptic commercial license API
  version: "1.0.0"
  contact:
    url: __DISTRIBUTION_CONTACT_URL__
  x-agentskeptic-distribution:
    repository: __DISTRIBUTION_REPO_URL__
    npmPackage: __DISTRIBUTION_NPM_URL__
    openApi: __OPENAPI_SELF_URL__
  description: "__IDENTITY_ONE_LINER__\n\nMachine-readable contract for license preflight used by the published npm CLI.\nBase URL is your deployed app origin (same as NEXT_PUBLIC_APP_URL).\n\nEvery path in this document returns an `x-request-id` response header on all status codes (echo a valid client `x-request-id` when supplied; otherwise server-generated). Non-2xx JSON bodies follow RFC 7807-style Problem Details (`type`, `title`, `status`, `detail`, optional `code`, `instance`) unless noted; `POST /api/v1/usage/reserve` denials additionally include legacy `allowed`, `code`, `message`, and optional `upgrade_url` for backward compatibility.\n"
...
servers:
  - url: __SERVERS_ORIGIN__
    description: Replace with your deployment origin
paths:
...
  /api/oss/claim-redeem:
    post:
      operationId: postOssClaimRedeem
      summary: Redeem OSS claim ticket (browser session; optional claim_secret in body)
      description: Requires an authenticated session; prefers pending cookie from GET /verify/link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OssClaimRedeemRequest"
      responses:
        "200":
          description: Redeemed; returns run summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OssClaimRedeemOk"
        "400":
          description: Claim failed or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Already linked to another account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/oss/claim-continuation:
    post:
      operationId: postOssClaimContinuation
      summary: Record browser-open acknowledgement for interactive-human tickets (CLI only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OssClaimContinuationRequest"
      responses:
        "204":
          description: Recorded or already recorded
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Not applicable or forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
...
    OssClaimRedeemRequest:
      type: object
      properties:
        claim_secret:
          type: string
          pattern: "^[0-9a-f]{64}$"
      additionalProperties: true
    OssClaimRedeemOk:
      type: object
      required:
        - schema_version
        - run_id
        - terminal_status
        - workload_class
        - subcommand
        - build_profile
        - claimed_at
      properties:
        schema_version:
          type: integer
          const: 1
        run_id:
          type: string
        terminal_status:
          type: string
        workload_class:
          type: string
        subcommand:
          type: string
        build_profile:
          type: string
        claimed_at:
          type: string
          format: date-time
    OssClaimContinuationRequest:
      type: object
      required: [claim_secret]
      properties:
        claim_secret:
          type: string
          pattern: "^[0-9a-f]{64}$"
      additionalProperties: false
...